pipeline/requirements.txt
httpx
numpy
openai
opuslib
pydub
PyYAML
scipy
webrtcvad
